Rename dynamo.py -> dynamodb.py
