+++ cogeo_mosaic/backends/__init__.py
@@ -1,5 +1,5 @@
 from .auto import auto_backend
-from .dynamo import DynamoDBBackend
+from .dynamodb import DynamoDBBackend
 from .file import FileBackend
 from .http import HttpBackend
 from .s3 import S3Backend
